requirements.txt
requests
beautifulsoup4
google-generativeai
pydantic
python-dotenv
